# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RoomFlash merchant and storefront API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerSegment } from './customerSegment';

export interface Customer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email: string | null;
  city: string;
  ordersCount: number;
  totalSpent: number;
  /** @nullable */
  lastOrderAt: string | null;
  segment: CustomerSegment;
}