# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RoomFlash merchant and storefront API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface DashboardSummary {
  revenue: number;
  orders: number;
  customers: number;
  averageOrderValue: number;
  /** @nullable */
  revenueChange: number | null;
  freeOrdersRemaining: number;
  orderLimit: number;
  lowStockCount: number;
  latestOrders: Order[];
}