# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RoomFlash merchant and storefront API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  number: string;
  customerName: string;
  customerPhone: string;
  customerCity: string;
  total: number;
  itemsCount: number;
  status: OrderStatus;
  paymentMethod: OrderPaymentMethod;
  createdAt: string;
}