# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RoomFlash merchant and storefront API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputCountry } from './orderInputCountry';
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  /** @minLength 2 */
  customerName: string;
  /** @minLength 5 */
  customerPhone: string;
  /** @nullable */
  email?: string | null;
  /** @minLength 5 */
  address: string;
  /** @minLength 2 */
  city: string;
  country: OrderInputCountry;
  /** @nullable */
  notes?: string | null;
  /** @minItems 1 */
  items: OrderItemInput[];
}